// backend/src/prisma/contract.d.ts
export type CodecTypes = PgTypes;
// ...
export namespace Models {
  export type public_Usuario = {
    id: CodecTypes['pg/int4@1']['output'];
    nome: CodecTypes['pg/text@1']['output'];
    email: CodecTypes['pg/text@1']['output'];
    senha: CodecTypes['pg/text@1']['output'];
    tipo: CodecTypes['pg/text@1']['output'];
    createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    readonly [RelationKeys]?: never;
  };
}